Add radio buttons to quiz. Utilized correct id and name structure.
Enabled navigation within page using HTML nav element. The items in the UL can be navigated to with an anchor linking to the id of the h2 elements. I uust use # to target the specific ID, like in CSS. 

Used [method="post"] & [action="link] to submit the form

Used role attribute to indicate the purpose behind an element. In this case, I used role="region" to aide in accessibility

Used [aria-labelleby] for section elements to improve accessibility.

Labels can detect the nearest value. There is no need to wrap labels around elements. 

Placeholders for inputs are usually not the best practice. Rely on using the text for labels instead.

Used span elements to improve accessibility for each question.

To change the text displayed on the submit button, use value="THE TEXT"

Use <br> to do a line break