Create accessibility functions to quiz
Used aspect ratio to maintain size of an image

Used Child > Combinator to select elements that are direct children of a specific parent. 

This rule below helps screen readers. This is a class that allows only screen readers to read. 
.sr-only {
  position: absolute;
  width: 1px;
  height: 1px;
  padding: 0;
  margin: -1px;
  overflow: hidden;
  clip: rect(0, 0, 0, 0);
  white-space: nowrap;
  border: 0;
}

Used :before to target the h3 element to give each the # symbol